requirements.txt
packaging
keras~=3.3.3

PyYAML~=6.0.1
